- impl feedbacks parsing and displaying in the main list - impl logout

+++ app/src/main/java/app/survey/android/feedbackapp/model/MainSurvey.java
@@ -5,10 +5,10 @@ import java.io.Serializable;
 public class MainSurvey implements Serializable {
     private int iconRes;
     private String title;
-    private int completedCount;
-    private int unsentCount;
+    private long completedCount;
+    private long unsentCount;
 
-    public MainSurvey(int iconRes, String title, int completedCount, int unsentCount) {
+    public MainSurvey(int iconRes, String title, long completedCount, long unsentCount) {
         this.iconRes = iconRes;
         this.title = title;
         this.completedCount = completedCount;
@@ -23,11 +23,11 @@ public class MainSurvey implements Serializable {
         return title;
     }
 
-    public int getCompletedCount() {
+    public long getCompletedCount() {
         return completedCount;
     }
 
-    public int getUnsentCount() {
+    public long getUnsentCount() {
         return unsentCount;
     }
 }
